+++ src/decorators/GraphQLResolver.ts
@@ -1,11 +1,17 @@
 import { MetadataStorage } from "../metadata/metadata-storage";
-import { ClassType } from "../types/decorators";
+import { ClassType, ClassTypeResolver } from "../types/decorators";
 
-export function GraphQLResolver(typeFunc: () => ClassType): ClassDecorator {
+export function GraphQLResolver(typeFunc: ClassTypeResolver): ClassDecorator;
+export function GraphQLResolver(objectType: ClassType): ClassDecorator;
+export function GraphQLResolver(objectTypeOrTypeFunc: Function): ClassDecorator {
+  const getParentType = objectTypeOrTypeFunc.prototype
+    ? () => objectTypeOrTypeFunc as ClassType
+    : (objectTypeOrTypeFunc as ClassTypeResolver);
+
   return target => {
     MetadataStorage.registerResolver({
       target,
-      getParentType: typeFunc,
+      getParentType,
     });
   };
 }