+++ tests/functional/description.ts
@@ -67,7 +67,7 @@ describe("Description", () => {
         describedField: string;
       }
 
-      @GraphQLResolver(() => SampleObject)
+      @GraphQLResolver(objectType => SampleObject)
       class SampleResolver {
         @Query()
         normalQuery(): string {